Added ala-file node
