+++ node-red-contrib-ala-file.html
@@ -1,0 +1,63 @@
+<script type="text/x-red" data-template-name="ala-file out">
+    <div class="form-row">
+        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+
+    <div class="form-row">
+        <label for="node-input-filename"><i class="icon-tag"></i> File name</label>
+        <input type="text" id="node-input-filename" placeholder="file name...">
+    </div>
+
+    <div class="form-row">
+        <label for="node-input-format"><i class="fa fa-code"></i> Format</label>
+        <select id="node-input-format" style="width:180px;">
+            <option value="xlsx">XLSX</option>
+            <option value="xls">XLS</option>
+            <option value="json">JSON</option>
+            <option value="txt">TXT</option>
+            <option value="csv">CSV</option>
+            <option value="tsv">TSV</option>
+            <option value="sql">SQL</option>
+        </select>
+    </div>
+
+    <div class="form-row">
+        <label for="node-input-columns"><i class="icon-tag"></i> Columns</label>
+        <input type="text" id="node-input-columns" placeholder="Select columns...">
+    </div>
+
+    <div class="form-row">
+        <label for="node-input-headers"><i class="fa fa-list"></i> Headers</label>
+        <input type="checkbox" id="node-input-headers" style="width:20px">
+    </div>
+
+</script>
+
+<script type="text/x-red" data-help-name="ala-file out">
+    <p>Writes <code>msg.payload</code> to the specified file in XLSX, CSV, or other formats. The node uses <a href="https://github.com/agershun/alasql" target="_blank">AlaSQL</a> library.</p>
+</script>
+
+<script type="text/javascript">
+    RED.nodes.registerType('ala-file out',{
+        category: 'advanced',
+        color:"#f7df5e",
+        inputs:1,
+        outputs:0,
+        defaults: {
+            name: {value:''},
+            filename: {value:''},
+            format: {value:'xlsx'},
+            columns: {value:'*'},
+            headers: {value:true}
+        },
+        icon: "alasql.png",
+        label: function() {
+            console.log(this.format);
+            return this.name||this.format||"ala file";
+        },
+        labelStyle: function() {
+            return this.name?"node_label_italic":"";
+        },
+    });
+</script>
