--- node-red-contrib-alasql-excel.html
@@ -1,50 +1,0 @@
-<!-- =========== EXCEL IN =============== -->
-
-<script type="text/x-red" data-template-name="excel in">
-    <div class="form-row">
-        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-
-    <div class="form-row">
-        <label for="node-input-filename"><i class="icon-tag"></i> File name</label>
-        <input type="text" id="node-input-filename" placeholder="file name...">
-    </div>
-
-</script>
-
-<script type="text/x-red" data-help-name="excel in">
-    <p>A Node-RED node to use <a href="https://github.com/agershun/alasql" target="_blank">AlaSQL</a> for fast SQL based in-memory data processing for BI and ERP applications.</p>
-
-</script>
-
-<script type="text/javascript">
-    RED.nodes.registerType('excel in',{
-        category: 'advanced',
-        color:"#f7df3e",
-        inputs:1,
-        outputs:0,
-        defaults: {
-            name: {value:''},
-            filename: {value:''}
-        },
-        icon: "alasql.png",
-        label: function() {
-            return this.name||"excel";
-        },
-        labelStyle: function() {
-            return this.name?"node_label_italic":"";
-        },
-        oneditprepare: function() {
-            // this.editor = RED.editor.createEditor({
-            //     id: 'node-input-query-editor',
-            //     mode: 'ace/mode/sql',
-            //     value: $("#node-input-query").val()
-            // });
-        },
-        oneditsave: function() {
-            // $("#node-input-query").val(this.editor.getValue())
-            // delete this.editor;
-        }
-    });
-</script>
